backend/requirements/base.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
gunicorn==20.1.0
pytz==2022.4
sqlparse==0.4.3
psycopg2-binary==2.9.5
python-dotenv==1.0.1
requests==2.32.3
responses==0.25.6